perf tests: Create structure for shell tests
The general tructure of test suites with tescases was implemented for C
tests for some time, while shell tests were just a list of testcases
without any possibility of grouping them together.

Provide the possibility of test suite structure for shell tests, when
the suite is created if the testcases are located in a subdirectory. All
of the deeper levels of directories will be merged with the first level
of test cases as the test stucute provides only a single level of
structuring. The naming of the test suite for not is given by the name
of the direcroty.