perf test: Provide setup for the shell test suite
As some of hte test cases require setup that would have to be done for
each testcase create the possibility to have a setup file that is run
before the test suite. The setup file will also have the possibility to
bear the name of the testsuite in case the dir name is not good enough.