perf test: Remove perftool drivers
The perf now provides all of the features required for running the
perftool test cases, such as creating log directories, running setup scripts
and the tests are structured by the base_ directories.

Remove the drivers as they are no longer necessary together with
the condition of skipping the base_ directories and run the
test cases by the default perf test structure.

Signed-off-by: Veronika Molnarova <vmolnaro@redhat.com>